refactor: add AuthForm, implement register
